requirements.txt
 fastapi==0.115.0 
 uvicorn==0.30.6 
 sqlalchemy==2.0.35 
 psycopg2-binary==2.9.9 
 alembic==1.13.3 
 pydantic==2.9.2 
 redis==5.0.8 
 xxhash==3.5.0 
 gunicorn==23.0.0 
 python-dotenv==1.0.1
 pydantic_settings>=2.0
 pybase62==0.4.3
